Подготовить проверку работоспособности Module Federation с profile
